fix(doctor), do not exit on errors
